add authorization for question editing/deleting

+++ app/Http/Controllers/QuestionsController.php
@@ -65,6 +65,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied. You can't edit a question from another user");
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -77,6 +81,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied. You can't edit a question from another user");
+        }
+
         $question->update($request->only('body', 'title'));
 
         return redirect('/questions')->with('success', 'Your question has been updated');
@@ -90,6 +98,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied. You can't delete a question from another user");
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
